The two main areas of conflict were changes to the platform
recorder API and changes to the mac native layer.

+++ toolchain/lc-compile/src/syntax-gen.c
@@ -45,12 +45,12 @@ enum SyntaxNodeKind
 typedef struct SyntaxNode *SyntaxNodeRef;
 struct SyntaxNodeMark
 {
-    long index;
+    intptr_t index;
     SyntaxNodeKind kind;
     int used;
     SyntaxNodeRef value;
-    long lmode;
-    long rmode;
+    intptr_t lmode;
+    intptr_t rmode;
 };
 
 struct SyntaxNode
@@ -76,24 +76,24 @@ struct SyntaxNode
         } keyword;
         struct
         {
-            long index;
+            intptr_t index;
             NameRef rule;
-            long lmode;
-            long rmode;
+            intptr_t lmode;
+            intptr_t rmode;
         } descent;
         struct
         {
-            long index;
-            long value;
+            intptr_t index;
+            intptr_t value;
         } boolean_mark, integer_mark;
         struct
         {
-            long index;
+            intptr_t index;
             double value;
         } real_mark;
         struct
         {
-            long index;
+            intptr_t index;
             NameRef value;
         } string_mark;
     };
@@ -104,8 +104,8 @@ struct SyntaxNode
     int mark_count;
     
     int left_trimmed, right_trimmed;
-    long left_lmode, left_rmode;
-    long right_lmode, right_rmode;
+    intptr_t left_lmode, left_rmode;
+    intptr_t right_lmode, right_rmode;
 };
 
 typedef enum SyntaxMethodType SyntaxMethodType;
@@ -121,8 +121,8 @@ typedef struct SyntaxArgument *SyntaxArgumentRef;
 struct SyntaxArgument
 {
     SyntaxArgumentRef next;
-    long mode;
-    long index;
+    intptr_t mode;
+    intptr_t index;
 };
 
 typedef struct SyntaxMethod *SyntaxMethodRef;
@@ -156,10 +156,10 @@ struct SyntaxRule
     NameRef module;
     NameRef name;
     SyntaxRuleKind kind;
-    long precedence;
+    intptr_t precedence;
     SyntaxNodeRef expr;
     SyntaxMethodRef methods;
-    long *mapping;
+    intptr_t *mapping;
     const char *deprecation_message;
     
     SyntaxMethodRef current_method;
@@ -178,8 +178,8 @@ static SyntaxRuleRef s_rule = NULL;
 static SyntaxNodeRef s_stack[MAX_NODE_DEPTH];
 static unsigned int s_stack_index = 0;
 
-static long *s_invoke_lists = NULL;
-static long s_invoke_list_size = 0;
+static intptr_t *s_invoke_lists = NULL;
+static intptr_t s_invoke_list_size = 0;
 
 static NameRef *s_unreserved_keywords = NULL;
 static unsigned int s_unreserved_keyword_count = 0;
@@ -190,7 +190,7 @@ static FILE* s_output;
 
 ////////////////////////////////////////////////////////////////////////////////
 
-static void BeginSyntaxRule(NameRef p_module, NameRef p_name, SyntaxRuleKind p_kind, long p_precedence)
+static void BeginSyntaxRule(NameRef p_module, NameRef p_name, SyntaxRuleKind p_kind, intptr_t p_precedence)
 {
     assert(s_rule == NULL);
     
@@ -226,27 +226,27 @@ void BeginExpressionSyntaxRule(NameRef p_module, NameRef p_name)
     BeginSyntaxRule(p_module, p_name, kSyntaxRuleKindExpression, 0);
 }
 
-void BeginPrefixOperatorSyntaxRule(NameRef p_module, NameRef p_name, long p_precedence)
+void BeginPrefixOperatorSyntaxRule(NameRef p_module, NameRef p_name, intptr_t p_precedence)
 {
     BeginSyntaxRule(p_module, p_name, kSyntaxRuleKindPrefixOperator, p_precedence);
 }
 
-void BeginPostfixOperatorSyntaxRule(NameRef p_module, NameRef p_name, long p_precedence)
+void BeginPostfixOperatorSyntaxRule(NameRef p_module, NameRef p_name, intptr_t p_precedence)
 {
     BeginSyntaxRule(p_module, p_name, kSyntaxRuleKindPostfixOperator, p_precedence);
 }
 
-void BeginLeftBinaryOperatorSyntaxRule(NameRef p_module, NameRef p_name, long p_precedence)
+void BeginLeftBinaryOperatorSyntaxRule(NameRef p_module, NameRef p_name, intptr_t p_precedence)
 {
     BeginSyntaxRule(p_module, p_name, kSyntaxRuleKindLeftBinaryOperator, p_precedence);
 }
 
-void BeginRightBinaryOperatorSyntaxRule(NameRef p_module, NameRef p_name, long p_precedence)
+void BeginRightBinaryOperatorSyntaxRule(NameRef p_module, NameRef p_name, intptr_t p_precedence)
 {
     BeginSyntaxRule(p_module, p_name, kSyntaxRuleKindRightBinaryOperator, p_precedence);
 }
 
-void BeginNeutralBinaryOperatorSyntaxRule(NameRef p_module, NameRef p_name, long p_precedence)
+void BeginNeutralBinaryOperatorSyntaxRule(NameRef p_module, NameRef p_name, intptr_t p_precedence)
 {
     BeginSyntaxRule(p_module, p_name, kSyntaxRuleKindNeutralBinaryOperator, p_precedence);
 }
@@ -576,9 +576,9 @@ static void JoinSyntaxNodes(SyntaxNodeRef p_left, SyntaxNodeRef p_right)
     FreeSyntaxNode(p_right);
 }
 
-static long CountSyntaxNodeMarks(SyntaxNodeRef p_node)
+static intptr_t CountSyntaxNodeMarks(SyntaxNodeRef p_node)
 {
-    long t_index = 0;
+    intptr_t t_index = 0;
     switch(p_node -> kind)
     {
         case kSyntaxNodeKindEmpty:
@@ -605,7 +605,7 @@ static long CountSyntaxNodeMarks(SyntaxNodeRef p_node)
 			t_index = 0;
             for(i = 0; i < p_node -> concatenate . operand_count; i++)
             {
-                long t_node_index;
+                intptr_t t_node_index;
                 t_node_index = CountSyntaxNodeMarks(p_node -> concatenate . operands[i]);
                 if (t_node_index > t_index)
                     t_index = t_node_index;
@@ -702,7 +702,7 @@ static void PrintSyntaxNode(SyntaxNodeRef p_node)
     }
 }
 
-static void RemoveFirstTermFromSyntaxNode(SyntaxNodeRef p_node, long *r_lmode, long *r_rmode)
+static void RemoveFirstTermFromSyntaxNode(SyntaxNodeRef p_node, intptr_t *r_lmode, intptr_t *r_rmode)
 {
     if (p_node -> kind == kSyntaxNodeKindAlternate)
     {
@@ -730,7 +730,7 @@ static void RemoveFirstTermFromSyntaxNode(SyntaxNodeRef p_node, long *r_lmode, l
         assert(0);
 }
 
-static void RemoveLastTermFromSyntaxNode(SyntaxNodeRef p_node, long *r_lmode, long *r_rmode)
+static void RemoveLastTermFromSyntaxNode(SyntaxNodeRef p_node, intptr_t *r_lmode, intptr_t *r_rmode)
 {
     if (p_node -> kind == kSyntaxNodeKindAlternate)
     {
@@ -912,7 +912,7 @@ void PushUnreservedKeywordSyntaxGrammar(const char *p_keyword)
     PushSyntaxNode(t_node);
 }
 
-void PushMarkedDescentSyntaxGrammar(long p_mark, NameRef p_rule, long p_lmode, long p_rmode)
+void PushMarkedDescentSyntaxGrammar(intptr_t p_mark, NameRef p_rule, intptr_t p_lmode, intptr_t p_rmode)
 {
     SyntaxNodeRef t_node;
     MakeSyntaxNode(&t_node);
@@ -934,7 +934,7 @@ void PushDescentSyntaxGrammar(NameRef p_rule)
     PushSyntaxNode(t_node);
 }
 
-void PushMarkedTrueSyntaxGrammar(long p_mark)
+void PushMarkedTrueSyntaxGrammar(intptr_t p_mark)
 {
     SyntaxNodeRef t_node;
     MakeSyntaxNode(&t_node);
@@ -944,7 +944,7 @@ void PushMarkedTrueSyntaxGrammar(long p_mark)
     PushSyntaxNode(t_node);
 }
 
-void PushMarkedFalseSyntaxGrammar(long p_mark)
+void PushMarkedFalseSyntaxGrammar(intptr_t p_mark)
 {
     SyntaxNodeRef t_node;
     MakeSyntaxNode(&t_node);
@@ -954,7 +954,7 @@ void PushMarkedFalseSyntaxGrammar(long p_mark)
     PushSyntaxNode(t_node);
 }
 
-void PushMarkedIntegerSyntaxGrammar(long p_mark, long p_value)
+void PushMarkedIntegerSyntaxGrammar(intptr_t p_mark, intptr_t p_value)
 {
     SyntaxNodeRef t_node;
     MakeSyntaxNode(&t_node);
@@ -964,7 +964,7 @@ void PushMarkedIntegerSyntaxGrammar(long p_mark, long p_value)
     PushSyntaxNode(t_node);
 }
 
-void PushMarkedRealSyntaxGrammar(long p_mark, long p_value)
+void PushMarkedRealSyntaxGrammar(intptr_t p_mark, intptr_t p_value)
 {
     SyntaxNodeRef t_node;
     MakeSyntaxNode(&t_node);
@@ -974,7 +974,7 @@ void PushMarkedRealSyntaxGrammar(long p_mark, long p_value)
     PushSyntaxNode(t_node);
 }
 
-void PushMarkedStringSyntaxGrammar(long p_mark, const char *p_value)
+void PushMarkedStringSyntaxGrammar(intptr_t p_mark, const char *p_value)
 {
     NameRef t_name;
 	SyntaxNodeRef t_node;
@@ -1055,7 +1055,7 @@ void PushFalseArgumentSyntaxMapping(void)
 {
 }
 
-void PushIntegerArgumentSyntaxMapping(long p_value)
+void PushIntegerArgumentSyntaxMapping(intptr_t p_value)
 {
 }
 
@@ -1067,7 +1067,7 @@ void PushStringArgumentSyntaxMapping(const char *p_value)
 {
 }
 
-static void PushMarkArgumentSyntaxMapping(long p_mode, long p_index)
+static void PushMarkArgumentSyntaxMapping(intptr_t p_mode, intptr_t p_index)
 {
     SyntaxArgumentRef t_arg;
     t_arg = (SyntaxArgumentRef)Allocate(sizeof(struct SyntaxArgument));
@@ -1085,30 +1085,30 @@ static void PushMarkArgumentSyntaxMapping(long p_mode, long p_index)
     }
 }
 
-void PushInMarkArgumentSyntaxMapping(long p_index)
+void PushInMarkArgumentSyntaxMapping(intptr_t p_index)
 {
     PushMarkArgumentSyntaxMapping(0, p_index);
 }
 
-void PushOutMarkArgumentSyntaxMapping(long p_index)
+void PushOutMarkArgumentSyntaxMapping(intptr_t p_index)
 {
     PushMarkArgumentSyntaxMapping(1, p_index);
 }
 
-void PushInOutMarkArgumentSyntaxMapping(long p_index)
+void PushInOutMarkArgumentSyntaxMapping(intptr_t p_index)
 {
     PushMarkArgumentSyntaxMapping(2, p_index);
 }
 
 ////////////////////////////////////////////////////////////////////////////////
 
-void AddUnreservedSyntaxKeyword(long p_name)
+void AddUnreservedSyntaxKeyword(intptr_t p_name)
 {
     s_unreserved_keywords = Reallocate(s_unreserved_keywords, sizeof(NameRef) * (s_unreserved_keyword_count + 1));
     s_unreserved_keywords[s_unreserved_keyword_count++] = (NameRef)p_name;
 }
 
-static void ListSyntaxNodeUnreservedKeywords(SyntaxNodeRef p_node, NameRef** x_tokens, long *x_token_count)
+static void ListSyntaxNodeUnreservedKeywords(SyntaxNodeRef p_node, NameRef** x_tokens, intptr_t *x_token_count)
 {
     switch(p_node -> kind)
     {
@@ -1147,7 +1147,7 @@ static void ListSyntaxNodeUnreservedKeywords(SyntaxNodeRef p_node, NameRef** x_t
     }
 }
 
-static void MergeSyntaxNodes(SyntaxNodeRef p_node, SyntaxNodeRef p_other_node, long *x_next_mark, long *x_mapping)
+static void MergeSyntaxNodes(SyntaxNodeRef p_node, SyntaxNodeRef p_other_node, intptr_t *x_next_mark, intptr_t *x_mapping)
 {
     if (p_node -> kind == kSyntaxNodeKindAlternate)
     {
@@ -1191,20 +1191,20 @@ static void MergeSyntaxNodes(SyntaxNodeRef p_node, SyntaxNodeRef p_other_node, l
 
 static void MergeSyntaxRule(SyntaxRuleRef p_rule, SyntaxRuleRef p_other_rule)
 {
-    long t_rule_mark_count, t_other_rule_mark_count;
-	long *t_mark_mapping;
+    intptr_t t_rule_mark_count, t_other_rule_mark_count;
+	intptr_t *t_mark_mapping;
 
     t_rule_mark_count = CountSyntaxNodeMarks(p_rule -> expr);
     t_other_rule_mark_count = CountSyntaxNodeMarks(p_other_rule -> expr);
     
-    t_mark_mapping = (long *)Allocate(t_other_rule_mark_count * sizeof(long));
+    t_mark_mapping = (intptr_t *)Allocate(t_other_rule_mark_count * sizeof(intptr_t));
     
     MergeSyntaxNodes(p_rule -> expr, p_other_rule -> expr, &t_rule_mark_count, t_mark_mapping);
     
     p_other_rule -> mapping = t_mark_mapping;
 }
 
-static void SetSyntaxNodeMarkAsUsed(struct SyntaxNodeMark *p_marks, int p_mark_count, long p_index, SyntaxNodeRef p_value)
+static void SetSyntaxNodeMarkAsUsed(struct SyntaxNodeMark *p_marks, int p_mark_count, intptr_t p_index, SyntaxNodeRef p_value)
 {
     int i;
 	for(i = 0; i < p_mark_count; i++)
@@ -1279,7 +1279,7 @@ static void GenerateSyntaxRuleTerm(SyntaxNodeRef p_node, struct SyntaxNodeMark *
     }
 }
 
-static void AddSyntaxNodeMark(struct SyntaxNodeMark **x_marks, int *x_mark_count, long p_index, SyntaxNodeKind p_kind, long p_lmode, long p_rmode)
+static void AddSyntaxNodeMark(struct SyntaxNodeMark **x_marks, int *x_mark_count, intptr_t p_index, SyntaxNodeKind p_kind, intptr_t p_lmode, intptr_t p_rmode)
 {
     int i;
 	for(i = 0; i < *x_mark_count; i++)
@@ -1393,7 +1393,7 @@ static void GenerateSyntaxRuleExplicitAndUnusedMarks(SyntaxNodeRef p_node)
                     if (p_node -> marks[i] . value -> integer_mark . value < 0)
                         fprintf(s_output, "EXPRESSION'integer(UndefinedPosition, %ld)", p_node -> marks[i] . value -> integer_mark . value);
                     else
-                        fprintf(s_output, "EXPRESSION'unsignedinteger(UndefinedPosition, %lu)", (unsigned long)p_node -> marks[i] . value -> integer_mark . value);
+                        fprintf(s_output, "EXPRESSION'unsignedinteger(UndefinedPosition, %lu)", (uintptr_t)p_node -> marks[i] . value -> integer_mark . value);
                     break;
                 case kSyntaxNodeKindRealMark:
                     fprintf(s_output, "EXPRESSION'real(UndefinedPosition, Mark%ldValue)", p_node -> marks[i] . index);
@@ -1721,9 +1721,9 @@ static void GenerateInvokeLists(void)
 static void GenerateTokenList(void)
 {
     NameRef *t_tokens;
-    long t_token_count;
+    intptr_t t_token_count;
 	SyntaxRuleGroupRef t_group;
-	long i;
+	intptr_t i;
 	unsigned int j;
 	const char *t_string;
 
@@ -1758,7 +1758,7 @@ static void GenerateTokenList(void)
     }
 }
 
-void DefineCustomInvokeList(long index, long ptr)
+void DefineCustomInvokeList(intptr_t index, intptr_t ptr)
 {
     if (index >= s_invoke_list_size)
     {
@@ -1769,12 +1769,12 @@ void DefineCustomInvokeList(long index, long ptr)
             s_invoke_list_size *= 2;
         }
     
-        s_invoke_lists = (long *)Reallocate(s_invoke_lists, s_invoke_list_size * sizeof(long));
+        s_invoke_lists = (intptr_t *)Reallocate(s_invoke_lists, s_invoke_list_size * sizeof(intptr_t));
     }
     s_invoke_lists[index] = ptr;
 }
 
-void LookupCustomInvokeList(long index, long *r_ptr)
+void LookupCustomInvokeList(intptr_t index, intptr_t *r_ptr)
 {
     *r_ptr = s_invoke_lists[index];
 }
